Unicode problem in graph json is resolved temporarily.
The problem was that we lose some quotes in json (representing graph etc.) somehow in the process while handling node-spec parameters.

Beware that from this point on, we do not support unicode strings. This is a quick fix; improve this later on. Solve it internally in the parameter handling parts...